Introduce flowRunner
