+++ .golangci.yml
@@ -36,8 +36,6 @@ linters:
     - dogsled
     - dupl
     - exportloopref
-    - gochecknoglobals
-    - gochecknoinits
     - gocognit
     - goconst
     - gocritic
@@ -74,6 +72,8 @@ linters:
     # - lll
     # - scopelint
     # - prealloc
+    # - gochecknoinits
+    # - gochecknoglobals
 
 issues:
   # enable issues excluded by default
